fix DESCRIPTION
See here for instructions:

https://www.bioconductor.org/help/faq/#developer-faq

Generally, we do not use "Depends:" except for rare circumstances.

The right place for dependencies is "Imports:", but we need to be
explicit about which functions we use from which packages.

For example, notice how I put `@importFrom yaml read_yaml` to indicate
that the `RunCRE_HSAStringDB()` function uses the `read_yaml()` function
from the `yaml` package.

I deleted 'dplyr' because it is only used in commented code, but it is
never used elsewhere.